+++ test/integration/examples_test.go
@@ -5,11 +5,14 @@ package integration
 
 import (
 	"bytes"
+	"context"
 	"fmt"
+	"net"
 	"net/http"
 	"os"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
 	"github.com/miekg/dns"
@@ -72,6 +75,41 @@ func TestHTTPRouteExample(t *testing.T) {
 			return strings.Contains(b.String(), "<title>httpbin.org</title>")
 		}
 		return false
+	}, ingressWait, waitTick)
+}
+
+var udpRouteExampleManifests = fmt.Sprintf("%s/gateway-udproute.yaml", examplesDIR)
+
+func TestUDPRouteExample(t *testing.T) {
+	t.Logf("applying yaml manifest %s", strings.TrimPrefix(udpRouteExampleManifests, examplesDIR))
+	b, err := os.ReadFile(udpRouteExampleManifests)
+	// TODO as of 2022-04-01, UDPRoute does not support using a different inbound port than the outbound
+	// destination service port. Once parentRef port functionality is stable, we should remove this
+	s := string(b)
+	s = strings.ReplaceAll(s, "port: 53", "port: 9999")
+	b = []byte(s)
+	require.NoError(t, err)
+	require.NoError(t, clusters.ApplyYAML(ctx, env.Cluster(), string(b)))
+
+	defer func() {
+		require.NoError(t, clusters.DeleteYAML(ctx, env.Cluster(), string(b)))
+	}()
+
+	t.Logf("configuring test and setting up API clients")
+	resolver := &net.Resolver{
+		PreferGo: true,
+		Dial: func(ctx context.Context, network, address string) (net.Conn, error) {
+			d := net.Dialer{
+				Timeout: time.Second * 5,
+			}
+			return d.DialContext(ctx, network, fmt.Sprintf("%s:%d", proxyUDPURL.Hostname(), 9999))
+		},
+	}
+
+	t.Logf("verifying that the UDPRoute becomes routable")
+	require.Eventually(t, func() bool {
+		_, err := resolver.LookupHost(ctx, "kernel.org")
+		return err == nil
 	}, ingressWait, waitTick)
 }